Replace binding creators with plugins
