+++ src/main/js/controller/binding-creators/input.ts
@@ -4,11 +4,14 @@ import {TypeUtil} from '../../misc/type-util';
 import {Color, RgbaColorObject} from '../../model/color';
 import {Point2d, Point2dObject} from '../../model/point-2d';
 import {Target} from '../../model/target';
-import * as BooleanInputBindingControllerCreators from './boolean-input';
-import * as ColorInputBindingControllerCreators from './color-input';
-import * as NumberInputBindingControllerCreators from './number-input';
-import * as Point2dInputBindingControllerCreators from './point-2d-input';
-import * as StringInputBindingControllerCreators from './string-input';
+import {BooleanInputPlugin} from './boolean-input';
+import {NumberColorInputPlugin} from './color-number-input';
+import {ObjectColorInputPlugin} from './color-object-input';
+import {StringColorInputPlugin} from './color-string-input';
+import {createController, InputBindingPlugin} from './input-binding-plugin';
+import {NumberInputPlugin} from './number-input';
+import {Point2dInputPlugin} from './point-2d-input';
+import {StringInputPlugin} from './string-input';
 
 export type InputIn = boolean | number | string | Color | Point2d;
 export type InputEx =
@@ -38,16 +41,21 @@ export function create(
 	}
 
 	const bc = [
-		BooleanInputBindingControllerCreators.create,
-		ColorInputBindingControllerCreators.createWithNumber,
-		ColorInputBindingControllerCreators.createWithObject,
-		ColorInputBindingControllerCreators.createWithString,
-		NumberInputBindingControllerCreators.create,
-		StringInputBindingControllerCreators.create,
-		Point2dInputBindingControllerCreators.create,
+		BooleanInputPlugin,
+		NumberColorInputPlugin,
+		ObjectColorInputPlugin,
+		StringColorInputPlugin,
+		NumberInputPlugin,
+		StringInputPlugin,
+		Point2dInputPlugin,
 	].reduce(
-		(result, createBindingController) =>
-			result || createBindingController(document, target, params),
+		(result, plugin: InputBindingPlugin<any, any>) =>
+			result ||
+			createController(plugin, {
+				document: document,
+				target: target,
+				inputParams: params,
+			}),
 		null,
 	);
 	if (bc) {
